// packages/vnicore/src/data/types.ts
export type V5GAnimationType =
  | "idle"
  | "move"
  | "multi_move"
  | "fade"
  | "scale_up"
  | "scale_down"
  | "scale_in"
  | "scale_out"
  | "pop"
  | "bounce_jump"
  | "shake"
  | "blink"
  | "rotate"
  | "slide_in"
  | "slide_out"
  | "bounce_in"
  | "pulse"
  | "float"
  | "swing"
  | "particles"
  | "particle_twinkle"
  | "particle_wall"
  | "particle_combo"
  | "particle_stream"
  | "chaser_light"
  | "gather_particles"
  | "smoke_mist"
  | "energy_ring"
  | "slash_light"
  | "flame_flicker"
  | "wave_band"
  | "wave_distort"
  | "speed_lines"
  | "drift_fall"
  | "path_particles"
  | "shatter"
  | "glow"
  | "safe_glow"
  | "squash_stretch"
  | "card_carousel_3d";